added restriction to some pages

+++ app/Http/Controllers/MissionController.php
@@ -4,7 +4,9 @@ namespace App\Http\Controllers;
 
 use App\Models\Mission;
 use App\Models\Contact;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Yajra\DataTables\Facades\DataTables;
 
 class MissionController extends Controller
@@ -14,43 +16,50 @@ class MissionController extends Controller
      */
     public function index(Request $request)
     {
-        if ($request->ajax()) {
+        $currentUser = Auth::user();
+        
+        if ($currentUser->hasRole(User::ROLE_ADMIN) || $currentUser->hasRole(User::ROLE_SUPER_ADMIN)) {
+        
+            if ($request->ajax()) {
 
-            $data = Mission::with('contact')->latest()->get();
+                $data = Mission::with('contact')->latest()->get();
 
-            return DataTables::of($data)
-                    ->addIndexColumn()
-                    ->addColumn('action', function($row){
+                return DataTables::of($data)
+                        ->addIndexColumn()
+                        ->addColumn('action', function($row){
 
-                           $btn = '<a href="javascript:void(0)" data-bs-toggle="tooltip"  data-id="'.$row->id.'" title="Details" class="details btn btn-info btn-sm showMission"><i class="fa fa-eye text-white"></i></a>';
-                           $btn = $btn.'<a href="javascript:void(0)" data-bs-toggle="tooltip"  data-id="'.$row->id.'" title="Modifier" class="edit btn btn-warning btn-sm editMission"><i class="fa fa-pencil text-white"></i></a>';
+                            $btn = '<a href="javascript:void(0)" data-bs-toggle="tooltip"  data-id="'.$row->id.'" title="Details" class="details btn btn-info btn-sm showMission"><i class="fa fa-eye text-white"></i></a>';
+                            $btn = $btn.' <a href="javascript:void(0)" data-bs-toggle="tooltip"  data-id="'.$row->id.'" title="Modifier" class="edit btn btn-warning btn-sm editMission"><i class="fa fa-pencil text-white"></i></a>';
 
-                           $btn = $btn.' <a href="javascript:void(0)" data-bs-toggle="tooltip" data-id="'.$row->id.'" title="Supprimer" class="btn btn-danger btn-sm deleteMission"><i class="fa fa-trash"></i></a>';
+                            $btn = $btn.' <a href="javascript:void(0)" data-bs-toggle="tooltip" data-id="'.$row->id.'" title="Supprimer" class="btn btn-danger btn-sm deleteMission"><i class="fa fa-trash"></i></a>';
 
-                            return $btn;
-                    })
-                    ->editColumn('nomMiss', function($row) {
-                        return ucfirst($row->nomMiss);
-                    })
-                    ->editColumn('adresse', function($row) {
-                        return ucfirst($row->contact->adresse);
-                    })
-                    ->editColumn('email', function($row) {
-                        return $row->contact->email;
-                    })
-                    ->editColumn('telMobile', function($row) {
-                        return $row->contact->telMobile;
-                    }) ->editColumn('created_at', function($row) {
-                        return date('d/m/Y H:i', strtotime($row->created_at));
-                    })
-                    ->editColumn('updated_at', function($row) {
-                        return date('d/m/Y H:i', strtotime($row->updated_at));
-                    })
-                    ->rawColumns(['action'])
-                    ->make(true);
+                                return $btn;
+                        })
+                        ->editColumn('nomMiss', function($row) {
+                            return ucfirst($row->nomMiss);
+                        })
+                        ->editColumn('adresse', function($row) {
+                            return ucfirst($row->contact->adresse);
+                        })
+                        ->editColumn('email', function($row) {
+                            return $row->contact->email;
+                        })
+                        ->editColumn('telMobile', function($row) {
+                            return $row->contact->telMobile;
+                        }) ->editColumn('created_at', function($row) {
+                            return date('d/m/Y H:i', strtotime($row->created_at));
+                        })
+                        ->editColumn('updated_at', function($row) {
+                            return date('d/m/Y H:i', strtotime($row->updated_at));
+                        })
+                        ->rawColumns(['action'])
+                        ->make(true);
+            }
+
+            return view('missions.show');
+        }else {
+            return redirect('dashboard');
         }
-
-        return view('missions.show');
     }
 
     /**
@@ -66,39 +75,44 @@ class MissionController extends Controller
      */
     public function store(Request $request)
     {
+        $currentUser = Auth::user();
+        
+        if ($currentUser->hasRole(User::ROLE_ADMIN) || $currentUser->hasRole(User::ROLE_SUPER_ADMIN)) {
+        
+            $contact = Contact::updateOrCreate([
 
-        $contact = Contact::updateOrCreate([
+                'id' => $request->contact_id
 
-            'id' => $request->contact_id
+            ],
 
-        ],
+            [
+                'adresse' => $request->adresse,
+                'email' => $request->email,
+                'telMobile' => $request->telMobile,
+                'telFixe' => $request->telFixe,
+                'BP' => $request->BP,
+                'codePost' => $request->codePost,
 
-        [
-            'adresse' => $request->adresse,
-            'email' => $request->email,
-            'telMobile' => $request->telMobile,
-            'telFixe' => $request->telFixe,
-            'BP' => $request->BP,
-            'codePost' => $request->codePost,
+            ]);
+            $mission = Mission::updateOrCreate([
 
-        ]);
-        $mission = Mission::updateOrCreate([
+                'id' => $request->mission_id
 
-            'id' => $request->mission_id
+            ],
 
-        ],
+            [
+                'nomMiss' => $request->nomMiss,
 
-        [
-            'nomMiss' => $request->nomMiss,
+            ]);
 
-        ]);
+            $mission->contact()->associate($contact);
 
-        $mission->contact()->associate($contact);
+            $mission->save();
 
-        $mission->save();
-
-    return response()->json(['success'=>'Mission enregistrée avec succès']);
-
+            return response()->json(['success'=>'Mission enregistrée avec succès']);
+        }else {
+            return redirect('dashboard');
+        }
     }
 
     /**
@@ -106,9 +120,15 @@ class MissionController extends Controller
      */
     public function show($id)
     {
-        $mission = Mission::with('contact')->find($id);
-        // dd($mission->contact);
-        return response()->json($mission);
+        $currentUser = Auth::user();
+        
+        if ($currentUser->hasRole(User::ROLE_ADMIN) || $currentUser->hasRole(User::ROLE_SUPER_ADMIN)) {
+            
+            $mission = Mission::with('contact')->find($id);
+            return response()->json($mission);
+        }else {
+            return redirect('dashboard');
+        }
     }
 
     /**
@@ -116,8 +136,15 @@ class MissionController extends Controller
      */
     public function edit($id)
     {
-        $mission = Mission::with('contact')->find($id);
-        return response()->json($mission);
+        $currentUser = Auth::user();
+            
+            if ($currentUser->hasRole(User::ROLE_ADMIN) || $currentUser->hasRole(User::ROLE_SUPER_ADMIN)) {
+            
+            $mission = Mission::with('contact')->find($id);
+            return response()->json($mission);
+        }else {
+            return redirect('dashboard');
+        }
     }
 
     /**
@@ -132,12 +159,18 @@ class MissionController extends Controller
      * Remove the specified resource from storage.
      */
     public function destroy($id)
-    {
-        $mission = Mission::find($id);
-        $contact = $mission->contact();
-        $mission->delete();
-        $contact->delete();
-        return response()->json(['success'=>'Mission supprimée avec succès.']);
-
+    {   
+        $currentUser = Auth::user();
+        
+        if ($currentUser->hasRole(User::ROLE_ADMIN) || $currentUser->hasRole(User::ROLE_SUPER_ADMIN)) {
+        
+            $mission = Mission::find($id);
+            $contact = $mission->contact();
+            $mission->delete();
+            $contact->delete();
+            return response()->json(['success'=>'Mission supprimée avec succès.']);
+        }else {
+            return redirect('dashboard');
+        }
     }
 }
