+++ app/Http/Controllers/TypeController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use App\Models\Type;
+use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Yajra\DataTables\Facades\DataTables;
 class TypeController extends Controller
 {
@@ -13,33 +15,40 @@ class TypeController extends Controller
     public function index(Request $request)
     {
 
-        if ($request->ajax()) {
+        $currentUser = Auth::user();
+        
+        if ($currentUser->hasRole(User::ROLE_ADMIN) || $currentUser->hasRole(User::ROLE_SUPER_ADMIN)) {
+        
+            if ($request->ajax()) {
 
-            $data = Type::latest()->get();
+                $data = Type::latest()->get();
 
-            return DataTables::of($data)
-                    ->addIndexColumn()
-                    ->addColumn('action', function($row){
+                return DataTables::of($data)
+                        ->addIndexColumn()
+                        ->addColumn('action', function($row){
 
-                           $btn = '<a href="javascript:void(0)" data-bs-toggle="tooltip"  data-id="'.$row->id.'" title="Modifier" class="edit btn btn-warning btn-sm editType"><i class="fa fa-pencil text-white"></i></a>';
+                            $btn = '<a href="javascript:void(0)" data-bs-toggle="tooltip"  data-id="'.$row->id.'" title="Modifier" class="edit btn btn-warning btn-sm editType"><i class="fa fa-pencil text-white"></i></a>';
 
-                           $btn = $btn.' <a href="javascript:void(0)" data-bs-toggle="tooltip" data-id="'.$row->id.'" title="Supprimer" class="btn btn-danger btn-sm deleteType"><i class="fa fa-trash"></i></a>';
+                            $btn = $btn.' <a href="javascript:void(0)" data-bs-toggle="tooltip" data-id="'.$row->id.'" title="Supprimer" class="btn btn-danger btn-sm deleteType"><i class="fa fa-trash"></i></a>';
 
-                            return $btn;
-                    })
-                    ->editColumn('libelleType', function($row) {
-                        return ucfirst($row->libelleType);
-                    }) ->editColumn('created_at', function($row) {
-                        return date('d/m/Y H:i', strtotime($row->created_at));
-                    })
-                    ->editColumn('updated_at', function($row) {
-                        return date('d/m/Y H:i', strtotime($row->updated_at));
-                    })
-                    ->rawColumns(['action'])
-                    ->make(true);
+                                return $btn;
+                        })
+                        ->editColumn('libelleType', function($row) {
+                            return ucfirst($row->libelleType);
+                        }) ->editColumn('created_at', function($row) {
+                            return date('d/m/Y H:i', strtotime($row->created_at));
+                        })
+                        ->editColumn('updated_at', function($row) {
+                            return date('d/m/Y H:i', strtotime($row->updated_at));
+                        })
+                        ->rawColumns(['action'])
+                        ->make(true);
+            }
+
+            return view('types.show');
+        }else {
+            return redirect('dashboard');
         }
-
-        return view('types.show');
     }
 
     /**
@@ -55,6 +64,10 @@ class TypeController extends Controller
      */
     public function store(Request $request)
     {
+        $currentUser = Auth::user();
+        
+        if ($currentUser->hasRole(User::ROLE_ADMIN) || $currentUser->hasRole(User::ROLE_SUPER_ADMIN)) {
+        
             Type::updateOrCreate([
 
                         'id' => $request->type_id
@@ -70,6 +83,9 @@ class TypeController extends Controller
 
 
             return response()->json(['success'=>'Statut enregistré avec succès']);
+        }else {
+            return redirect('dashboard');
+        }
     }
 
 
@@ -78,7 +94,13 @@ class TypeController extends Controller
      */
     public function show(Type $type)
     {
-        //
+        $currentUser = Auth::user();
+        
+        if ($currentUser->hasRole(User::ROLE_ADMIN) || $currentUser->hasRole(User::ROLE_SUPER_ADMIN)) {
+        
+        }else {
+            return redirect('dashboard');
+        }
     }
 
     /**
@@ -86,9 +108,16 @@ class TypeController extends Controller
      */
     public function edit($id)
     {
-        $type = Type::find($id);
+        $currentUser = Auth::user();
+        
+        if ($currentUser->hasRole(User::ROLE_ADMIN) || $currentUser->hasRole(User::ROLE_SUPER_ADMIN)) {
+            
+            $type = Type::find($id);
 
-        return response()->json($type);
+            return response()->json($type);
+        }else {
+            return redirect('dashboard');
+        }
     }
 
     /**
@@ -104,7 +133,14 @@ class TypeController extends Controller
      */
     public function destroy($id)
     {
-        Type::find($id)->delete();
-        return response()->json(['success'=>'Type supprimé avec succès.']);
+        $currentUser = Auth::user();
+        
+        if ($currentUser->hasRole(User::ROLE_ADMIN) || $currentUser->hasRole(User::ROLE_SUPER_ADMIN)) {
+            
+            Type::find($id)->delete();
+            return response()->json(['success'=>'Type supprimé avec succès.']);
+        }else {
+            return redirect('dashboard');
+        }
     }
 }
